// src/app/enum/roles.enum.ts
export enum UserRoles {
  SERVICE_PROVIDER_ADMIN = "SPA",
  SERVICE_PROVIDER_ADMIN_CLONE = "SPAC",
  CLINICAL_FACILITY_ADMIN = "CFA",
  CLINICAL_FACILITY_ADMIN_CLONE = "CFAC",
  GENERAL_CLINICIAN = "GC",
  SUPERVISORY_CLINICIAN = "SC",
  PHYSICIAN = "PHY",
  SUPER_ADMIN = "SA",
  CLINICAL_FACILITY_USER = "CFU",
}